requirements.txt
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
